assignments/A3/requirements.txt
pyenchant
nltk
